delimit entities consistently
Delimit `&nbsp` entity with `;` consistently in the file. (It is sometimes and then isn't other times. This makes it that way all the time.)

Yeah, this is the tiniest of changes. Feel free to ignore. Or not. You know, like this text here. You don't need to read it. Unless you want to.